+++ Plugin/CheckoutControllerQuoteReset.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * Yireo SalesBlock2 for Magento
  *
@@ -8,18 +8,17 @@
  * @license     Open Source License (OSL v3)
  */
 
-declare(strict_types=1);
-
 namespace Yireo\SalesBlock2\Plugin;
 
 use Magento\Checkout\Controller\Onepage as OnepageController;
-use Magento\Checkout\Model\Cart;
-use Magento\Checkout\Model\Session as CheckoutSession;
 use Magento\Framework\Controller\Result\RedirectFactory;
 use Magento\Framework\Controller\ResultInterface;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NotFoundException;
 use Magento\Framework\Message\ManagerInterface;
 use Yireo\SalesBlock2\Helper\Rule as RuleHelper;
+use Yireo\SalesBlock2\Logger\Debugger;
+use Yireo\SalesBlock2\Utils\DestroyQuote;
 
 /**
  * Plugin ControllerQuoteReset
@@ -33,11 +32,6 @@ class CheckoutControllerQuoteReset
     private $ruleHelper;
 
     /**
-     * @var CheckoutSession
-     */
-    private $checkoutSession;
-
-    /**
      * @var RedirectFactory
      */
     private $resultRedirectFactory;
@@ -48,36 +42,41 @@ class CheckoutControllerQuoteReset
     private $messageManager;
 
     /**
-     * @var Cart
+     * @var DestroyQuote
      */
-    private $cart;
+    private $destroyQuote;
+    /**
+     * @var Debugger
+     */
+    private $debugger;
 
     /**
      * PreventAddToCart constructor.
      * @param RuleHelper $ruleHelper
-     * @param CheckoutSession $checkoutSession
      * @param RedirectFactory $resultRedirectFactory
      * @param ManagerInterface $messageManager
-     * @param Cart $cart
+     * @param DestroyQuote $destroyQuote
+     * @param Debugger $debugger
      */
     public function __construct(
         RuleHelper $ruleHelper,
-        CheckoutSession $checkoutSession,
         RedirectFactory $resultRedirectFactory,
         ManagerInterface $messageManager,
-        Cart $cart
+        DestroyQuote $destroyQuote,
+        Debugger $debugger
     ) {
         $this->ruleHelper = $ruleHelper;
-        $this->checkoutSession = $checkoutSession;
         $this->resultRedirectFactory = $resultRedirectFactory;
         $this->messageManager = $messageManager;
-        $this->cart = $cart;
+        $this->destroyQuote = $destroyQuote;
+        $this->debugger = $debugger;
     }
 
     /**
      * @param OnepageController $subject
      * @param ResultInterface $result
      * @return ResultInterface
+     * @throws LocalizedException
      */
     public function afterExecute(
         OnepageController $subject,
@@ -86,19 +85,12 @@ class CheckoutControllerQuoteReset
         try {
             $match = $this->ruleHelper->findMatch();
             $this->messageManager->addWarningMessage($match->getMessage());
-
-            $quote = $this->checkoutSession->getQuote();
-            foreach ($quote->getAllVisibleItems() as $item) {
-                $itemId = $item->getItemId();
-                $this->cart->removeItem($itemId)->save();
-            }
-            $this->checkoutSession->clearStorage();
-            $this->checkoutSession->resetCheckout();
-
+            $this->destroyQuote->destroy();
             $resultRedirect = $this->resultRedirectFactory->create();
             $resultRedirect->setUrl('/checkout/cart');
             return $resultRedirect;
         } catch (NotFoundException $exception) {
+            $this->debugger->debug('Plugin for Magento\Checkout\Controller\Onepage: No match found');
             return $result;
         }
     }